core/local: Rename ./atom_watcher to ./atom/watcher
- Same namespace as everything else atom-related
- Easier completion (no `/` vs `_`)
