+++ core/local/atom/watcher.js
@@ -3,30 +3,30 @@
 const Promise = require('bluebird')
 const _ = require('lodash')
 
-const checksumer = require('./checksumer')
-const logger = require('../logger')
+const checksumer = require('./../checksumer')
+const logger = require('../../logger')
 
-const Producer = require('./atom/producer')
-const addInfos = require('./atom/add_infos')
-const filterIgnored = require('./atom/filter_ignored')
-const winDetectMove = require('./atom/win_detect_move')
-const winIdenticalRenaming = require('./atom/win_identical_renaming')
-const scanFolder = require('./atom/scan_folder')
-const awaitWriteFinish = require('./atom/await_write_finish')
-const initialDiff = require('./atom/initial_diff')
-const addChecksum = require('./atom/add_checksum')
-const incompleteFixer = require('./atom/incomplete_fixer')
-const overwrite = require('./atom/overwrite')
-const dispatch = require('./atom/dispatch')
+const Producer = require('./producer')
+const addInfos = require('./add_infos')
+const filterIgnored = require('./filter_ignored')
+const winDetectMove = require('./win_detect_move')
+const winIdenticalRenaming = require('./win_identical_renaming')
+const scanFolder = require('./scan_folder')
+const awaitWriteFinish = require('./await_write_finish')
+const initialDiff = require('./initial_diff')
+const addChecksum = require('./add_checksum')
+const incompleteFixer = require('./incomplete_fixer')
+const overwrite = require('./overwrite')
+const dispatch = require('./dispatch')
 
 /*::
-import type Pouch from '../pouch'
-import type Prep from '../prep'
+import type Pouch from '../../pouch'
+import type Prep from '../../prep'
 import type EventEmitter from 'events'
-import type { Ignore } from '../ignore'
-import type { Checksumer } from './checksumer'
-import type { AtomEventsDispatcher } from './atom/dispatch'
-import type { Scanner } from './atom/producer'
+import type { Ignore } from '../../ignore'
+import type { Checksumer } from '../checksumer'
+import type { AtomEventsDispatcher } from './dispatch'
+import type { Scanner } from './producer'
 
 type AtomWatcherOptions = {
   syncPath: string,
