+++ test/support/helpers/local.js
@@ -11,7 +11,7 @@ const conflictHelpers = require('./conflict')
 const { ContextDir } = require('./context_dir')
 
 const Local = require('../../../core/local')
-const atomWatcher = require('../../../core/local/atom_watcher')
+const atomWatcher = require('../../../core/local/atom/watcher')
 const { TMP_DIR_NAME } = require('../../../core/local/constants')
 const dispatch = require('../../../core/local/atom/dispatch')
 const { INITIAL_SCAN_DONE } = require('../../../core/local/atom/event')
