+++ test/unit/local/atom/watcher.js
@@ -3,13 +3,13 @@
 
 const should = require('should')
 
-const AtomWatcher = require('../../../core/local/atom_watcher')
-const initialDiff = require('../../../core/local/atom/initial_diff')
+const AtomWatcher = require('../../../../core/local/atom/watcher')
+const initialDiff = require('../../../../core/local/atom/initial_diff')
 
-const configHelpers = require('../../support/helpers/config')
-const pouchHelpers = require('../../support/helpers/pouch')
+const configHelpers = require('../../../support/helpers/config')
+const pouchHelpers = require('../../../support/helpers/pouch')
 
-describe('core/local/atom_watcher', () => {
+describe('core/local/atom/watcher', () => {
   before('instanciate config', configHelpers.createConfig)
   before('instanciate pouch', pouchHelpers.createDatabase)
   after('clean pouch', pouchHelpers.cleanDatabase)
